financial-doc-automation/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pymupdf==1.24.5
faiss-cpu==1.8.0
sentence-transformers==3.0.1
google-generativeai==0.7.2
python-dotenv==1.0.1
streamlit==1.36.0
numpy==1.26.4
requests==2.32.3
